Rules.copy: add ability to override transformer options
A user might want to override an option in the options
object which belongs to the transformer and is provided
as a parameter in all calls to all rule functions.

Obviously a custom made rule function can ignore the
options object entirely, but the rules.copy uses it
for the path separator.

The transformer-provided options can be overridden
by providing an options property on the params object
when creating the rule function.