Change package exports to be more convenient
There is now a default export of transformer.
The rules and parsers that may come later are all
supplemental, and for the sake of convenience. The
core functionality is all in the transformer, so it
makes sense for that to be the default export.

The rules are all exported as named exports so that
by default, they will have short names. This makes
a big difference when creating a ruleset.