Create integration test for transformer, rules, parsers etc.
The purpose of this is to make sure everything works
together as expected, and to test out the main exports
like a consumer of the package would use them.

The initial tests are to make sure it is possible
to re-implement the functionality from the url-que
ry-to-prisma package (which was really kind of a
proto version of this package, much more specific
and far less re-usable).