Implement and test parseDate convenience function
This is designed to be imported and used just like
a built in js parser like parseInt or parseFloat.
Normally when adding a parser parameter to the copy
function, it has to be written like:

{ parser: (v) => new Date(v) }

which is not that convenient. With parseDate, you
can just write:

{ parser: parseDate }

Like the rules file, parseDate is a named export.
This is to allow more parsers to be added later
if required, without messing up imports on projects
that already use parseDate.