transformer: options object now takes an optional nestedInputKeys property
If this property is set to false, then an input key
of 'my.input.key' will try to retrieve the value from:

{
  'my.input.key': 'my value'
}

instead of:

{
  my: {
    input: {
      key: 'my value'
    }
  }
}

A nestedOutputKeys property is also included on the
options object, to be passed to all the rules, which
will determine where the value ends up. It does not
affect anything internally on the object-transformer.