Revert to default deepMerge behaviour
The rationale here is that the default deepMerge
behaviour is the most complicated to implement -
putting both values in an array if they are different,
leaving a single copy if they are the same.

It is trivial to override this behaviour to overwrite
any existing values with a custom conflictHandler, but
not the other way round. Therefore, the copy function's
conflictHandler will default to undefined, and when
deepMerge receives undefined as a conflictHandler
parameter it will just do the default behaviour.

Additionally, this means by default no data will
be lost, unlike with an overwrite as the default.