Implement skipping empty strings per copy rule
E.g. if transformer is set to keep empty strings,
but you want one particular rule to omit, then
you can.

However if the transformer is set to omit empty
strings, it will totally skip that key's rule,
regardless of whether the copy rule was created
with omitEmptyStrings set to false.