Move implementation of skipping empty strings entirely from transformer to rule functions
Previously, this functionality was kind of split
between the transformer and the rule functions.

If options.omitEmptyStrings was true,
transformer would skip every empty value, even if
a rule had the local options.omitEmptyStrings set
to true - the rule would just be skipped entirely.

That meant that if transformer options.omitEmptyStrings
were false, the rule would run and in the copy rule,
you were able to override omitEmptyStrings, but not the
other way round. It was not consistent.

This way, whether empty strings are skipped or not
is entirely the remit of the rule, not the transformer,
and will be more flexible. The transformer will just
keep a default setting that can be set by the user,
which is passed to all rules as part of the options
object and can be overridden in copy, or just ignored
entirely in a custom rule.