JSDocs: make transformer and copy summaries cleaner
Previously where JSDocs typedefs and callbacks were
imported from jsdoc.common.mjs, the summary of each
function included the filepath instead of just saying
the type. This looked very cluttered and hard to read.

So each file that imports from jsdoc.common.mjs creates
its own local typedef with the same name. The summaries
are now a lot cleaner and easier to read for the end
user that is trying to use the package functions.