Throw an error if the transformer function is called without an inputObject
If a user tries to transform null, or an array,
or undefined, they probably made a mistake. It would
be better to point that out rather than ignore it,
since it may produce unpredictable results, or the
error may manifest in some other error message that
is harder to figure out.